fix:rework folder structure,separation business logic
