+++ src/auth/auth.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
+import { AuthController } from './api/auth.controller';
+import { AuthService } from './service/auth.service';
 import { UserModule } from 'src/user/user.module';
 import { TokenModule } from '../token/token.module';
