+++ src/user/service/user.service.ts
@@ -10,23 +10,14 @@ import { AppError } from 'common/errors/errors';
 export class UserService {
     constructor(@InjectModel(User) private readonly userRepository: UserRepository) {
     }
-    async deleteUser(userId: number): Promise<boolean> {
-        await this.userRepository.delete(userId);
-        return true;
-    }
     async hashPassword(password: string) {
         return bcrypt.hash(password, 10)
-    }
-    async findUserByUserName(userName: string) {
-        return this.userRepository.findOne({ where: { userName: userName } })
     }
     async createUser(dto: CreateUserDTO): Promise<User> {
         dto.password = await this.hashPassword(dto.password);
         const createdUser = this.userRepository.create(dto);
         await this.userRepository.save(createdUser);
         return createdUser;
-        console.log('a');
-
     }
 
 }
